% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{adjR2}
\alias{adjR2}
\title{Adjusted R-squared}
\usage{
adjR2(..., digits, verbose)
}
\arguments{
\item{...}{one of several model fit objects.}

\item{digits}{an (optional) integer value indicating the number of decimal places to be used.}

\item{verbose}{an (optional) logical indicating if should the report of results be printed.}
}
\value{
A matrix with the values of the adjusted R-squared for all model fit objects.
}
\description{
Computes the adjusted R-squared
}
