% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Advertising}
\format{
A data frame with 200 rows and 4 variables:
\describe{
  \item{TV}{a numeric vector indicating the advertising budget on TV.}
  \item{radio}{a numeric vector indicating the advertising budget on radio.}
  \item{newspaper}{a numeric vector indicating the advertising budget on newspaper.}
  \item{sales}{a numeric vector indicating the sales of the interest product.}
}
}
\source{
\url{https://www.statlearning.com/s/Advertising.csv}
}
\usage{
data(advertising)
}
\description{
The Advertising data set consists of the sales of that product in
200 different markets, along with advertising budgets for the product in each
of those markets for three different media: TV, radio, and newspaper.
}
\examples{
data(advertising)
pairs(~ sales + TV + radio + newspaper, pch=20, data = advertising)
}
\references{
James G., Witten D., Hastie T., Tibshirani R. (2013, page 15) \emph{An Introduction to Statistical Learning with Applications in R}, Springer, New York.
}
\keyword{datasets}
