% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{envelope}
\alias{envelope}
\title{Normal QQ-plot with simulated envelope of model residuals}
\usage{
envelope(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A matrix with the simulated envelope and, optionally, a plot of it.
}
\description{
Generic function for building a normal QQ-plot with simulated envelope of residuals obtained from a fitted model.
}
