% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{estequa}
\alias{estequa}
\title{Function to extract estimating equations}
\usage{
estequa(model, ...)
}
\arguments{
\item{model}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector with the value of the estimating equations evaluated at the parameter estimates and the observed data.
}
\description{
Extracts estimating equations evaluated at the parameter estimates and the observed data for a fitted model object.
}
