% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fabric}
\alias{fabric}
\title{Fabric faults}
\format{
A data frame with 32 rows and 2 variables:
\describe{
  \item{roll}{a numeric vector indicating the length of the rolls.}
  \item{faults}{a numeric vector indicating the number of faults.}
}
}
\usage{
data(fabric)
}
\description{
The main objective of the analysis of this dataset is to assess
if there is an association between the number of faults in rolls of fabric and
their length.
}
\examples{
data(fabric)
with(fabric,plot(roll, faults, pch=16, xlab="Length of roll", ylab="Number of faults"))
}
\references{
Hinde J., Demetrio C.G.B. (1998) Over-dispersion: models and estimation. \emph{Computational Statistics & Data Analysis} 27:151–170.
}
\keyword{datasets}
