% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{glance.glmgee}
\alias{glance.glmgee}
\title{Glance at a(n) glmgee object}
\usage{
\method{glance}{glmgee}(x, ...)
}
\arguments{
\item{x}{an object of class \emph{glmgee}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Glance summarizes information about a GEE model.
}
