% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{gvif}
\alias{gvif}
\title{Generalized Variance Inflation Factor}
\usage{
gvif(model, ...)
}
\arguments{
\item{model}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object with the values of the GVIF for all effects in the model.
}
\description{
Computes the generalized variance inflation factor (GVIF) for a fitted model object.
}
