% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{leverage}
\alias{leverage}
\title{Leverage}
\usage{
leverage(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object with the values of the leverage measures.
}
\description{
Computes leverage measures for a fitted model object.
}
