% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{localInfluence}
\alias{localInfluence}
\title{Local Influence}
\usage{
localInfluence(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object with the measures of local influence.
}
\description{
Computes measures of local influence for a fitted model object.
}
