% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Melanopus}
\alias{Melanopus}
\title{Assay of an Insecticide with a Synergist}
\format{
A data frame with 15 rows and 4 variables:
\describe{
  \item{Killed}{a numeric vector indicating how many grasshoppers were killed.}
  \item{Exposed}{a numeric vector indicating how many grasshoppers were exposed.}
  \item{Insecticide}{a numeric vector indicating the dose of insecticide.}
  \item{Synergist}{a numeric vector indicating the dose of synergist.}
}
}
\usage{
data(Melanopus)
}
\description{
These data are concerned with the estimation of the lowest-cost mixtures of insecticides and synergists. They
relate to assays on a grasshopper \emph{Melanopus sanguinipes} with carbofuran and piperonyl butoxide, which enhances
carbofuran's toxicity.
}
\references{
McCullagh P., Nelder J.A. (1989). \emph{Generalized Linear Models. 2nd Edition}. Chapman and Hall, London.

Wei B.C. (1998). \emph{Exponential Family Nonlinear Models}. Springer, Singapore.
}
\keyword{datasets}
