% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Oranges}
\alias{Oranges}
\title{Oranges}
\format{
A data frame with 35 rows and 3 variables:
\describe{
  \item{Trunk}{a numeric vector indicating the trunk circumference, in millimeters.}
  \item{Days}{a numeric vector indicating the time, in days, since December 31, 1968.}
  \item{Tree}{a numeric vector with the identifier of each orange tree.}
}
}
\usage{
data(Oranges)
}
\description{
The data arose from five orange trees grown at Riverside,
California, during the period 1969-1973. The response is the trunk
circumference, in millimeters, and the predictor variable is the time,
in days, with an arbitrary origin taken on December 31, 1968.
}
\examples{
dev.new()
data(Oranges)
with(Oranges,plot(Days, Trunk, pch=16, col="blue"))
}
\references{
Draper N., Smith H. (1998) Applied Regression Analysis, Third Edition. \emph{John Wiley & Sons}.
}
\keyword{datasets}
