% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{stepCriterion}
\alias{stepCriterion}
\title{Variable selection in regression models from a chosen criterion}
\usage{
stepCriterion(model, ...)
}
\arguments{
\item{model}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list which includes the descriptions of the linear predictors of the initial and final models as well as the criterion used to compare the candidate models.
}
\description{
Generic function for selecting variables from a fitted regression model using a chosen criterion.
}
