% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tidy.glmgee}
\alias{tidy.glmgee}
\title{Tidy a(n) glmgee object}
\usage{
\method{tidy}{glmgee}(
  x,
  conf.int = FALSE,
  conf.level = 0.95,
  exponentiate = FALSE,
  varest = c("robust", "df-adjusted", "bias-corrected", "model"),
  ...
)
}
\arguments{
\item{x}{an object of class \emph{glmgee}.}

\item{conf.int}{an (optional) character string indicating whether or not to include a confidence interval in the tidied output. As default, \code{conf.int} is set to \code{FALSE}.}

\item{conf.level}{an (optional) value indicating the confidence level to use for the confidence interval if \code{conf.int=TRUE}. As default, \code{conf.level} is set to 0.95.}

\item{exponentiate}{an (optional) logical indicating whether or not to exponentiate the coefficient estimates. As default, \code{exponentiate} is set to \code{FALSE}.}

\item{varest}{an (optional) character string indicating the type of estimator which should be used. The available options are: robust sandwich-type estimator ("robust"), degrees-of-freedom-adjusted estimator ("df-adjusted"), bias-corrected estimator ("bias-corrected"), and the model-based or naive estimator ("model"). As default, \code{type} is set to "robust".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Tidy summarizes information about the components of a GEE model.
}
