% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{vdtest}
\alias{vdtest}
\title{Test for Varying Dispersion Parameter}
\usage{
vdtest(model, ...)
}
\arguments{
\item{model}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list which includes the main attributes of the test as, for example, value of the statistic and \emph{p}-value.
}
\description{
Generic function for testing for varying dispersion parameter from a fitted model.
}
