% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm.R
\name{localInfluence.glmgee}
\alias{localInfluence.glmgee}
\title{Local Influence for Generalized Estimating Equations}
\usage{
\method{localInfluence}{glmgee}(
  object,
  type = c("total", "local"),
  perturbation = c("cw-clusters", "cw-observations", "response", "covariate"),
  covariate,
  coefs,
  plot.it = FALSE,
  identify,
  ...
)
}
\arguments{
\item{object}{an object of class glmgee obtained from the fit of a generalized estimating equation.}

\item{type}{an (optional) character string indicating the type of approach to study the local influence. The options are: the absolute value of the elements of the eigenvector which corresponds to the maximum absolute eigenvalue ("local"); and the absolute value of the elements of the main diagonal ("total"). By default, \code{type} is set to be "total".}

\item{perturbation}{an (optional) character string indicating the perturbation scheme to apply. The options are: case weight perturbation of clusters ("cw-clusters"); Case weight perturbation of observations ("cw-observations"); perturbation of covariates ("covariate"); and perturbation of response ("response"). By default, \code{perturbation} is set to be "cw-clusters".}

\item{covariate}{an character string which (partially) match with the names of one of the parameters in the linear predictor. This is only appropriate if \code{perturbation="covariate"}.}

\item{coefs}{an (optional) character string which (partially) match with the names of some of the parameters in the linear predictor.}

\item{plot.it}{an (optional) logical indicating if the plot of the measures of local influence is required or just the data matrix in which that plot is based. By default, \code{plot.it} is set to be FALSE.}

\item{identify}{an (optional) integer indicating the number of clusters/observations to identify on the plot of the measures of local influence. This is only appropriate if \code{plot.it=TRUE}.}

\item{...}{further arguments passed to or from other methods. If \code{plot.it=TRUE} then \code{...} may be used to include graphical parameters to customize the plot. For example, \code{col}, \code{pch}, \code{cex}, \code{main}, \code{sub}, \code{xlab}, \code{ylab}.}
}
\value{
A matrix as many rows as clusters/observations in the sample and one column with the values of the measures of local influence.
}
\description{
Computes some measures and, optionally, display	graphs of them to perform influence analysis based on the approaches described in Cook (1986) and Jung (2008).
}
\examples{
mod <- size ~ poly(days,4) + treat
fit <- glmgee(mod, id=tree, family=Gamma("log"), corstr="AR-1", data=spruces)
summary(fit)
localInfluence(fit,type="total",perturbation="cw-clusters",coefs="treat",plot.it=TRUE)

}
\references{
Cook D. (1986) Assessment of Local Influence. \emph{Journal of the Royal Statistical Society: Series B (Methodological)} 48, 133-155.

Jung K.-M. (2008) Local Influence in Generalized Estimating Equations. \emph{Scandinavian Journal of Statistics} 35, 286-294.
}
