% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glms.R
\name{residuals2}
\alias{residuals2}
\title{Residuals for Linear and Generalized Linear Models}
\usage{
residuals2(object, type, standardized = FALSE, plot.it = TRUE, identify, ...)
}
\arguments{
\item{object}{a object of the class lm or glm obtained from the fit of a linear or a generalized linear model.}

\item{type}{an (optional) character string giving the type of residuals which should be returned. The available options for LMs are: (1) externally studentized ("external"); (2) internally studentized ("internal") (default). The available options for GLMs are: (1) "pearson"; (2) "deviance";  (3) "quantile" (default).}

\item{standardized}{an (optional) logical switch indicating if the residuals should be standardized by dividing by the square root of \eqn{(1-h)}, where \eqn{h} is a measure of leverage. By default, \code{standardized} is set to be FALSE.}

\item{plot.it}{an (optional) logical switch indicating if a plot of the residuals is required. By default, \code{plot.it} is set to be TRUE.}

\item{identify}{an (optional) integer value indicating the number of individuals to identify on the plot of residuals. This is only appropriate when \code{plot.it=TRUE}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A vector with the observed residuals type \code{type}.
}
\description{
Computes residuals for a fitted linear or generalized linear model.
}
\examples{
# Example 1
fit1 <- lm(Species ~ Biomass + pH + Biomass*pH, data=richness)
residuals2(fit1, type="external", col="red", pch=20, col.lab="blue",
           col.axis="blue", col.main="black", family="mono", cex=0.8)

# Example 2
fit2 <- glm(infections ~ frequency + location, family=poisson, data=swimmers)
residuals2(fit2, type="quantile", col="red", pch=20,col.lab="blue",
           col.axis="blue",col.main="black",family="mono",cex=0.8)
}
